requirements.txt
numpy>=1.24
scipy>=1.10
torch>=2.2.0
torchvision>=0.17.0
timm>=0.9.12
transformers>=4.40.0
datasets>=2.18.0
pandas>=2.1.0
scikit-learn>=1.3.0
pyyaml>=6.0.0
wandb>=0.16.0
matplotlib>=3.8.0
seaborn>=0.12.0

